# [Compiler] Optimize compilation time by caching parsed modules